requirements.txt
pandas>=1.3.0
numpy>=1.21.0
psycopg2-binary>=2.9.0
python-dateutil>=2.8.0
